backend/requirements.txt
fastapi==0.115.9
uvicorn==0.34.2
python-multipart==0.0.20
PyYAML==6.0.2
python-jose[cryptography]==3.4.0
aiofiles==24.1.0
docker==7.1.0

